// src/payment/dto/verify-razorpay.dto.ts
import { IsNotEmpty, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class VerifyRazorpayDto {
  @ApiProperty({ example: 'order_9A33XWu170gUtm', description: 'Razorpay Order ID' })
  @IsString()
  @IsNotEmpty()
  razorpayOrderId!: string;

  @ApiProperty({ example: 'pay_2989dh82h', description: 'Razorpay Payment ID' })
  @IsString()
  @IsNotEmpty()
  razorpayPaymentId!: string;

  @ApiProperty({ example: '9ef4d79b23991a25786f8662', description: 'HMAC SHA256 signature generated by Razorpay client SDK' })
  @IsString()
  @IsNotEmpty()
  razorpaySignature!: string;
}